refactor(transform): split static resolver
