+++ scripts/check-production-ts-size.mjs
@@ -12,10 +12,6 @@ const LEGACY_OVERSIZED_FILES = new Map([
     'packages/transform/src/transform/generators/rewriteOxcBarrelImports.ts',
     1452,
   ],
-  [
-    'packages/transform/src/transform/generators/resolveStaticOxcValues.ts',
-    6084,
-  ],
   ['packages/transform/src/utils/collectOxcExportsAndImports.ts', 1436],
   ['packages/transform/src/utils/oxcPreevalTransforms.ts', 1794],
   ['packages/transform/src/utils/oxcShaker.ts', 1082],
